+++ shopify/packages/hybrid-cart-api/src/index.ts
@@ -51,7 +51,7 @@ export const createHybridCartApi = ({ debug = false }) => {
               error._tag === "RequestError" ||
               error._tag === "ResponseError" ||
               error._tag ===
-              "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
+                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
             ) {
               return Effect.fail(
                 new Error(error.message, {
@@ -93,7 +93,7 @@ export const createHybridCartApi = ({ debug = false }) => {
               error._tag === "RequestError" ||
               error._tag === "ResponseError" ||
               error._tag ===
-              "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
+                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
             ) {
               return Effect.fail(
                 new Error(error.message, {
@@ -135,7 +135,7 @@ export const createHybridCartApi = ({ debug = false }) => {
               error._tag === "RequestError" ||
               error._tag === "ResponseError" ||
               error._tag ===
-              "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
+                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
             ) {
               return Effect.fail(
                 new Error(error.message, {
@@ -178,7 +178,7 @@ export const createHybridCartApi = ({ debug = false }) => {
               error._tag === "RequestError" ||
               error._tag === "ResponseError" ||
               error._tag ===
-              "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
+                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
             ) {
               return Effect.fail(
                 new Error(error.message, {
@@ -221,7 +221,7 @@ export const createHybridCartApi = ({ debug = false }) => {
               error._tag === "RequestError" ||
               error._tag === "ResponseError" ||
               error._tag ===
-              "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
+                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
             ) {
               return Effect.fail(
                 new Error(error.message, {
@@ -266,9 +266,7 @@ export const createHybridCartApi = ({ debug = false }) => {
               if (
                 error._tag === "ParseError" ||
                 error._tag === "RequestError" ||
-                error._tag === "ResponseError" ||
-                error._tag ===
-                "@brytdesigns/hybrid-cart-client/InvalidAjaxMethodError"
+                error._tag === "ResponseError"
               ) {
                 return Effect.fail(
                   new Error(error.message, {