+++ shopify/utils/src/effect/Resource/data.ts
@@ -2,6 +2,7 @@ import type { ParseError } from "effect/ParseResult";
 
 import * as Effect from "effect/Effect";
 import * as Schema from "effect/Schema";
+import * as ParseResult from "effect/ParseResult";
 import { ID, GID, Type } from "./schema.js";
 
 type FormatOptions<T extends ID, R extends Type> = {
@@ -18,11 +19,26 @@ export const format = <const T extends ID, const R extends Type>({
   never
 > =>
   Effect.gen(function* () {
-    const actualId = yield* Schema.decode(ID)(id);
-    const actualType = yield* Schema.decode(Type)(type);
+    const [actualId, actualType] = yield* Effect.all(
+      [Schema.decode(ID)(id), Schema.decode(Type)(type)],
+      {
+        concurrency: "unbounded",
+      },
+    );
     return `gid://shopify/${actualType}/${actualId}` as `gid://shopify/${R}/${T}`;
   });
 
-export const parse = (
-  input: string,
-): Effect.Effect<string, ParseError, never> => Schema.decode(GID)(input);
+export const parse = (input: string) =>
+  Schema.decode(
+    Schema.transformOrFail(ID, GID, {
+      encode: (value, _, ast) =>
+        ParseResult.fail(
+          new ParseResult.Forbidden(
+            ast,
+            value,
+            "Encoding an ID to a GID is not allowed. Use the format function to convert IDs to GIDs.",
+          ),
+        ),
+      decode: (value) => ParseResult.succeed(`${value}`.split("/").pop()!),
+    }),
+  )(input);