+++ shopify/packages/ajax-cart-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-ajax-cart-api
 
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/shopify-utils@1.2.15
+
 ## 1.0.27
 
 ### Patch Changes